--- src/main/java/aero/cubox/errManagement/service/ErrManagementService.java
@@ -1,17 +1,0 @@
-package aero.cubox.errManagement.service;
-
-import aero.cubox.core.vo.FaceFeatureErrVO;
-
-import java.util.HashMap;
-import java.util.List;
-
-public interface ErrManagementService {
-
-    List<FaceFeatureErrVO> selectFaceFeatureErrList(FaceFeatureErrVO vo) throws Exception;
-
-    int getFaceFeatureErrCount(FaceFeatureErrVO vo) throws Exception;
-
-    FaceFeatureErrVO selectFaceFeatureErrOne(FaceFeatureErrVO vo) throws Exception;
-
-
-}
