+++ src/main/java/aero/cubox/report/service/impl/ReportDAO.java
@@ -1,9 +1,6 @@
 package aero.cubox.report.service.impl;
 
-import aero.cubox.core.vo.AlarmHistVO;
-import aero.cubox.core.vo.EntHistBioVO;
-import aero.cubox.core.vo.EntHistVO;
-import aero.cubox.core.vo.FaceVO;
+import aero.cubox.core.vo.*;
 import egovframework.rte.psl.dataaccess.EgovAbstractMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -49,5 +46,17 @@ public class ReportDAO extends EgovAbstractMapper {
 
     public FaceVO selectFaceOne(int faceId) throws Exception {
         return  selectOne ( "report.selectFaceOne", faceId);
+    }
+
+    public List<FaceFeatureErrVO> selectFaceFeatureErrList(FaceFeatureErrVO vo) throws Exception {
+        return selectList (sqlNameSpace + "selectFaceFeatureErrList", vo);
+    }
+
+    public int getFaceFeatureListCount(FaceFeatureErrVO vo) {
+        return selectOne(sqlNameSpace+"getFaceFeatureListCount", vo);
+    }
+
+    public FaceFeatureErrVO selectFaceFeatureErrOne(FaceFeatureErrVO vo) {
+        return selectOne(sqlNameSpace+"selectFaceFeatureErrOne", vo);
     }
 }
