+++ src/main/java/aero/cubox/report/service/impl/ReportServiceImpl.java
@@ -1,9 +1,6 @@
 package aero.cubox.report.service.impl;
 
-import aero.cubox.core.vo.AlarmHistVO;
-import aero.cubox.core.vo.EntHistBioVO;
-import aero.cubox.core.vo.EntHistVO;
-import aero.cubox.core.vo.FaceVO;
+import aero.cubox.core.vo.*;
 import aero.cubox.report.service.ReportService;
 import egovframework.rte.fdl.cmmn.EgovAbstractServiceImpl;
 import org.springframework.stereotype.Service;
@@ -57,6 +54,20 @@ public class ReportServiceImpl extends EgovAbstractServiceImpl implements Report
     @Override
     public FaceVO selectFaceOne(int faceId) throws Exception {
         return reportDAO.selectFaceOne(faceId);
+    }
+
+    @Override
+    public List<FaceFeatureErrVO> selectFaceFeatureErrList(FaceFeatureErrVO vo) throws Exception {
+        return reportDAO.selectFaceFeatureErrList(vo);
+    }
+
+    public int getFaceFeatureErrCount(FaceFeatureErrVO vo) throws Exception {
+        return reportDAO.getFaceFeatureListCount(vo);
+    }
+
+    @Override
+    public FaceFeatureErrVO selectFaceFeatureErrOne(FaceFeatureErrVO vo) throws Exception {
+        return reportDAO.selectFaceFeatureErrOne(vo);
     }
 
 }
