+++ src/main/resources/egovframework/sqlmap/cubox/Report.xml
@@ -177,5 +177,57 @@
         where id = #{faceId}
     </select>
 
+    <!-- 특징점추출 오류 목록 조회 -->
+    <select id="selectFaceFeatureErrList" parameterType="aero.cubox.core.vo.FaceFeatureErrVO" resultType="aero.cubox.core.vo.FaceFeatureErrVO">
+        SELECT
+        e.id, f.emp_cd, face_state_typ, c1.cd_nm as face_state_typ_nm,
+        face_feature_typ, c2.cd_nm as face_feature_typ_nm, e.error, e.created_at,
+        emp.emp_nm, emp.dept_nm
+        FROM T_FACE_FEATURE_ERR e
+        JOIN T_FACE f on f.id = e.face_id
+        LEFT OUTER JOIN T_EMP emp on f.emp_cd = emp.emp_cd
+        LEFT OUTER JOIN T_CMMN_CD c1 on c1.cd = f.face_state_typ and c1.cd_typ = 'FaceStateTyp'
+        LEFT OUTER JOIN T_CMMN_CD c2 on c2.cd = e.face_feature_typ and c2.cd_typ = 'FaceFeatureTyp'
+        WHERE 1=1
+        <if test='@aero.cubox.util.CommonUtils@notEmpty(emp_cd)'>
+            AND f.emp_cd = #{emp_cd}
+        </if>
+        <if test='@aero.cubox.util.CommonUtils@notEmpty(face_feature_typ)'>
+            AND e.face_feature_typ = #{face_feature_typ}
+        </if>
+        <if test='@aero.cubox.util.CommonUtils@notEmpty(emp_nm)'>
+            AND emp.emp_nm LIKE CONCAT('%',#{emp_nm},'%')
+        </if>
+        order by e.created_at desc
+        limit	#{srchCnt} offset #{offset}
+    </select>
+
+
+    <select id="getFaceFeatureListCount" parameterType="aero.cubox.core.vo.FaceFeatureErrVO" resultType="java.lang.Integer">
+        SELECT count(1)
+        FROM T_FACE_FEATURE_ERR e
+        JOIN T_FACE f on f.id = e.face_id
+        LEFT OUTER JOIN T_EMP emp on f.emp_cd = emp.emp_cd
+        WHERE 1=1
+        <if test='@aero.cubox.util.CommonUtils@notEmpty(emp_cd)'>
+            AND f.emp_cd = #{emp_cd}
+        </if>
+        <if test='@aero.cubox.util.CommonUtils@notEmpty(face_feature_typ)'>
+            AND e.face_feature_typ = #{face_feature_typ}
+        </if>
+        <if test='@aero.cubox.util.CommonUtils@notEmpty(emp_nm)'>
+            AND emp.emp_nm LIKE CONCAT('%',#{emp_nm},'%')
+        </if>
+    </select>
+
+    <!-- 특징점추출 오류 목록 조회 -->
+    <select id="selectFaceFeatureErrOne" parameterType="aero.cubox.core.vo.FaceFeatureErrVO" resultType="aero.cubox.core.vo.FaceFeatureErrVO">
+        SELECT
+            f.face_img
+        FROM T_FACE_FEATURE_ERR e
+                 JOIN T_FACE f on f.id = e.face_id
+        WHERE e.id = #{id}
+    </select>
+
 
 </mapper>
